Bledy do rozwiazania

+++ Item.cpp
@@ -2,5 +2,5 @@
 
 
 std::size_t Item::getPrice() const{
-    return static_cast<size_t>(rarity_) * price_;
+    return static_cast<int>(rarity_) * price_;
 }
